refactored checkCommentOwnership middleware
