+++ middleware/index.js
@@ -1,4 +1,5 @@
-const Campground = require('../models/campground');
+const Campground = require('../models/campground'),
+      Comment    = require('../models/comment');
 
 function isLoggedIn(req,res,next){
     if(req.isAuthenticated()){
@@ -28,4 +29,24 @@ function checkCampgroundOwnership(req,res,next){
     }
 }
 
-module.exports = {isLoggedIn, checkCampgroundOwnership};
+function checkCommentOwnership(req,res,next){
+    if(req.isAuthenticated()){
+        Comment.findById(req.params.comments_id,function(err,foundComment){
+            if(err){
+                res.redirect("back");
+            }else{
+                if(foundComment.author.id.equals(req.user._id)){     // using .equals because first variable is an
+                                                                        // object while other is string so == or ===
+                                                                        // will not work
+                    next();
+                }else{
+                    res.redirect("back");
+                }
+            }
+        })
+    }else{
+        res.redirect("back");
+    }
+}
+
+module.exports = {isLoggedIn, checkCampgroundOwnership, checkCommentOwnership};
