+++ routes/comments.js
@@ -3,12 +3,14 @@ var express=require("express"),
     Campground=require("../models/campground"),
     Comment=require("../models/comment");
 
+const middleware = require('../middleware');
+
 // 
 // COMMENTS ROUTES
 // 
 
 // Comments new
-router.get("/new", isLoggedIn, function(req,res){
+router.get("/new", middleware.isLoggedIn, function(req,res){
     Campground.findById(req.params.id,function(err,campground){
         if(err){
             console.log(err);
@@ -19,7 +21,7 @@ router.get("/new", isLoggedIn, function(req,res){
 });
 
 // Comments create
-router.post("/", isLoggedIn, function(req,res){
+router.post("/", middleware.isLoggedIn, function(req,res){
 
     // lookup campground using id
     Campground.findById(req.params.id,function(err,campground){
@@ -52,7 +54,7 @@ router.post("/", isLoggedIn, function(req,res){
 });
 
 // COMMENT EDIT ROUTE
-router.get("/:comments_id/edit", checkCommentOwnership, function(req,res){
+router.get("/:comments_id/edit", middleware.checkCommentOwnership, function(req,res){
     Comment.findById(req.params.comments_id,function(err,foundComment){
         if(err){
             console.log(err);
@@ -65,7 +67,7 @@ router.get("/:comments_id/edit", checkCommentOwnership, function(req,res){
 
 
 // COMMENT UPDATE ROUTE
-router.put("/:comments_id", checkCommentOwnership, function(req,res){
+router.put("/:comments_id", middleware.checkCommentOwnership, function(req,res){
     Comment.findByIdAndUpdate(req.params.comments_id, req.body.comment,function(err,updatedComment){
         if(err){
             console.log(err);
@@ -77,7 +79,7 @@ router.put("/:comments_id", checkCommentOwnership, function(req,res){
 })
 
 // COMMENT DELETE ROUTE
-router.delete("/:comments_id", checkCommentOwnership, function(req,res){
+router.delete("/:comments_id", middleware.checkCommentOwnership, function(req,res){
     Comment.findByIdAndRemove(req.params.comments_id,function(err){
         if(err){
             console.log(err);
@@ -86,36 +88,5 @@ router.delete("/:comments_id", checkCommentOwnership, function(req,res){
         res.redirect("/campgrounds/"+req.params.id);
     })
 })
-
-
-function checkCommentOwnership(req,res,next){
-    if(req.isAuthenticated()){
-        Comment.findById(req.params.comments_id,function(err,foundComment){
-            if(err){
-                res.redirect("back");
-            }else{
-                if(foundComment.author.id.equals(req.user._id)){     // using .equals because first variable is an
-                                                                        // object while other is string so == or ===
-                                                                        // will not work
-                    next();
-                }else{
-                    res.redirect("back");
-                }
-            }
-        })
-    }else{
-        res.redirect("back");
-    }
-}
-
-
-// middleware
-function isLoggedIn(req,res,next){
-    if(req.isAuthenticated()){
-        return next();
-    }else{
-        res.redirect("/login");
-    }
-}
 
 module.exports=router;
