+++ routes/index.js
@@ -10,7 +10,6 @@ router.get("/",function(req,res){
 });
 
 
-
 // AUTH ROUTES
 
 // show register form
@@ -59,15 +58,5 @@ router.get("/logout",function(req,res){
     req.logout();
     res.redirect("/campgrounds");
 })
-
-
-// middleware
-function isLoggedIn(req,res,next){
-    if(req.isAuthenticated()){
-        return next();
-    }else{
-        res.redirect("/login");
-    }
-}
 
 module.exports=router;
